[SITES] Improved search for terrain objects to hide.
`nearestTerrainObjects` searches for object centrepoints within a specified area.

This is problematic as it will not exclude larger objects like rocks or buildings whose centrepoint is not within the area, but they extend over into the area.

We need to search for objects whose area intersects with the site's area, and then remove them if they are present.

So... bounding boxes on the terrain objects!