CaptureTheFlag: Fixups for the CTF task loop and add some optional buildables
Due to a mismatch in variable locality, the CTF task was not running because the `holdAction` for dac cong players was not visible.
The `canLower` variable was attached to the flag object server-side, but `holdAction`s are attached to players, operating within player namespace locality.

Now, we set a single `publicVariable "vn_mf_bn_dc_target_flag"` which broadcasts the target flag to all players.
THe `holdAction`s should now work.

Also included in this change:

- [TWEAKED] Flag lowering/raising during defend phase -- multiple attempts can be chained together. Total `holdAction` time over all attempts is 10 seconds.
- [ADDED] Optional FOB building tasks during capture phase -- Helo Landing Pad and a Latrine.
- [FIX] Searching for completed optional buildables -- non player built objects no longer count.
- [FIX] Searching for completed optional buildables -- search area for flags was error'ing out due to faulty marker variable lookup.